--- src/main/java/shop/sajotuna/order/common/rabbitmq/FatalMessageRepository.java
@@ -1,6 +1,0 @@
-package shop.sajotuna.order.common.rabbitmq;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FatalMessageRepository extends JpaRepository<FatalMessageLog, Long> {
-}
