+++ src/main/java/shop/sajotuna/order/coupon/rabbitmq/CouponRabbitMqConfig.java
@@ -1,0 +1,28 @@
+package shop.sajotuna.order.coupon.rabbitmq;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.amqp.rabbit.config.SimpleRabbitListenerContainerFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.boot.autoconfigure.amqp.SimpleRabbitListenerContainerFactoryConfigurer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@RequiredArgsConstructor
+@Configuration
+public class CouponRabbitMqConfig {
+
+    @Bean("couponListenerContainerFactory")
+    public SimpleRabbitListenerContainerFactory couponListenerContainerFactory(
+            ConnectionFactory connectionFactory,
+            SimpleRabbitListenerContainerFactoryConfigurer configurer) {
+
+        SimpleRabbitListenerContainerFactory factory = new SimpleRabbitListenerContainerFactory();
+        configurer.configure(factory, connectionFactory);
+
+        factory.setConcurrentConsumers(5);
+        factory.setMaxConcurrentConsumers(10);
+        factory.setPrefetchCount(10);
+
+        return factory;
+    }
+}
