+++ src/main/java/shop/sajotuna/order/coupon/rabbitmq/CouponRabbitProperties.java
@@ -1,0 +1,4 @@
+package shop.sajotuna.order.coupon.rabbitmq;
+
+public class CouponRabbitProperties {
+}
