+++ src/main/java/shop/sajotuna/order/point/rabbitmq/PointFatalMessageRepository.java
@@ -1,0 +1,6 @@
+package shop.sajotuna.order.point.rabbitmq;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PointFatalMessageRepository extends JpaRepository<PointFatalMessageLog, Long> {
+}
