+++ src/main/java/shop/sajotuna/order/point/service/impl/PointQueueServiceImpl.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
 import shop.sajotuna.order.common.domain.Money;
-import shop.sajotuna.order.common.rabbitmq.PointRabbitProperties;
+import shop.sajotuna.order.point.rabbitmq.PointRabbitProperties;
 import shop.sajotuna.order.point.domain.PointPolicyType;
 import shop.sajotuna.order.point.service.dto.event.PointEvent;
 import shop.sajotuna.order.point.service.PointQueueService;
